Issue #1775 move com.ibm.ws.session.store.common to its own bundle
