+++ dev/com.ibm.websphere.appserver.sessionCache-1.0/bnd.bnd
@@ -16,6 +16,7 @@
 -dependson: \
     com.ibm.websphere.javaee.jcache.1.1;version=latest, \
 	com.ibm.websphere.security;version=latest, \
-	com.ibm.ws.session.db;version=latest, \
 	com.ibm.ws.serialization;version=latest, \
+	com.ibm.ws.session.db;version=latest, \
+	com.ibm.ws.session.store;version=latest, \
 	com.ibm.ws.session;version=latest
