+++ dev/com.ibm.websphere.appserver.sessionDatabase-1.0/bnd.bnd
@@ -14,6 +14,7 @@
 
 -dependson: \
 	com.ibm.websphere.security;version=latest, \
-	com.ibm.ws.session.db;version=latest, \
 	com.ibm.ws.serialization;version=latest, \
+	com.ibm.ws.session.db;version=latest, \
+	com.ibm.ws.session.store;version=latest, \
 	com.ibm.ws.session;version=latest
