+++ dev/com.ibm.ws.session.store/.gitignore
@@ -1,0 +1,2 @@
+/bin/
+/generated/
