+++ dev/com.ibm.ws.session.store/.settings/org.eclipse.core.resources.prefs
@@ -1,0 +1,4 @@
+#Ant properties
+#Automatically generated by the ant prepare.settings.files task
+eclipse.preferences.version=1
+encoding/<project>=UTF-8
