+++ dev/com.ibm.ws.session.store/build.xml
@@ -1,0 +1,18 @@
+<!--
+    Copyright (c) 2018 IBM Corporation and others.
+    All rights reserved. This program and the accompanying materials
+    are made available under the terms of the Eclipse Public License v1.0
+    which accompanies this distribution, and is available at
+    http://www.eclipse.org/legal/epl-v10.html
+   
+    Contributors:
+        IBM Corporation - initial API and implementation
+ -->
+<project name="com.ibm.ws.session.store" default="build">
+  <import file="../ant_build/public_imports/standard_imports.xml"/>
+
+  <!-- Disable trace/ffdc injection as it is hard-coded in this project --> 
+  <property name="no.trace.instrumentation" value="true"/>
+	<!-- Disable autoformatting as this code needs to be diff'ed with WAS code -->
+	<property name="no.auto.format" value="true"/>
+</project>
