+++ dev/com.ibm.ws.session.store/src/com/ibm/ws/session/store/common/internal/J2EEObjectWrapper.java
@@ -8,7 +8,7 @@
  * Contributors:
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
-package com.ibm.ws.session.store.common;
+package com.ibm.ws.session.store.common.internal;
 
 import java.io.Serializable;
